Rename service.rs to cache.rs, since it Service is not impld
