+++ tower-ready-cache/src/lib.rs
@@ -1,4 +1,4 @@
+pub mod cache;
 pub mod error;
-pub mod service;
 
-pub use self::service::ReadyCache;
+pub use self::cache::ReadyCache;
